fix missing localization of "Layout" menu.
the reason was that _after_ loading the current language from the persistence,
new GUI (localizable) elements were constructed, but the whole implementation of all of them currently
constructs them un-localized. their localization only occurs on an explicit language change.

so the fix simply is to after construction of ALL GUI elements finally notify them of a language "change" to the current language.