modified the standard DockItems behaviour: the default now is that they are really CLOSED (i.e., removed) upon closure.
previously, they only got "hidden".
they that way accumulated unwantedly in the config.

we really want to close them.

of course, before closure
- they are asked to save their persistency
- they are asked if they agree to closure, they still can cancel that

to achieve this effect, the previous flag "CloseOnHide" has been removed.

a new flag has been added "PreventClosing". if a component sets THAT flag, a close will only be a "hide".
the component will stay alive invisibly.
this for example might be necessary for some infrastructure windows.