Introduced new function IsCloseOK() for components.
They will be asked by the ComponentManager when they are about to be closed,
for example when the user closes its window, or the user shuts down the application.
In case the component has unsaved content, e.g., an edited document,
it should prompt the user with a messagebox "Do you want to save XYZ? Yes/No/Cancel".
On Cancel, it should return false from IsCloseOK(). In that case, the closing will be aborted.

If this function returns true, the component then will be asked to save its persistency by calling Save(),
and after that, its Closed() function will be invoked.
There, you can cleanup everything that must be cleaned up before the destructor runs,
e.g. stopping internal threads etc.
MAYBE in future we can abolish the Closed() function and let the destructor of components do all that work.