Workaround the problem that the build process locally copied GTKsharp runtime DLLs into the build products folder.
This is definitely unwanted because the scenario
"build on Windows, but run the build product on Linux"
must be supported.

We definitely want to catch the system-wide installed GTKsharp runtime libraries.

This can be achieved by setting the <Private>False</Private> option in the *.csproj files,
which also can be edited in the Visual Studio GUI, the switch is there named "Local Copy" with inversed boolean semantics.

Despite now setting this option everywhere (which was a tedious manual editing of *.csproj files........),
still Visual Studio for some reason copies such files there.
We could not completely get rid of them.

We had no other choice than to add a small buildprocess workaround batch which EXPLICITLY after the build checks for the
existence of the unwanted DLLs, and, if present, deletes them.

Jira-Issue: TEMPO-639
Change-Id: I9eadf0cfcd3d2910eaa2a8e0cede9207bda63469