deactivate strange code which produces null object Exceptions on Ubuntu 18.04
It is not clear to me what this code tries to achieve.
It seemingly relies on a specific GLib version to be present,
and accesses its guts by some strange direct calling.
The code was only used by the 2 classes "DockContainer" and "DockFrame".

The code says it wants to workaround some memory leaking.
So, with its removal, the leaks could come back, but they anyhow need
to be fixed differently IMHO.